+++ eventsourcing/application/multiprocess.py
@@ -9,7 +9,7 @@ from eventsourcing.application.simple import ApplicationWithConcreteInfrastructu
 from eventsourcing.application.system import DEFAULT_POLL_INTERVAL, PromptOutbox, System, SystemRunner
 from eventsourcing.domain.model.decorators import retry
 from eventsourcing.domain.model.events import subscribe, unsubscribe
-from eventsourcing.exceptions import CausalDependencyFailed, OperationalError, RecordConflictError
+from eventsourcing.exceptions import CausalDependencyFailed, OperationalError, RecordConflictError, ProgrammingError
 from eventsourcing.infrastructure.base import DEFAULT_PIPELINE_ID
 
 
@@ -55,6 +55,8 @@ class MultiprocessRunner(SystemRunner):
         for pipeline_id in self.pipeline_ids:
             for process_name, upstream_names in self.system.followings.items():
                 process_class = self.system.process_classes[process_name]
+                inbox = self.inboxes[(pipeline_id, process_name.lower())]
+                outbox = self.outboxes.get((pipeline_id, process_name.lower()))
                 os_process = OperatingSystemProcess(
                     application_process_class=process_class,
                     infrastructure_class=self.infrastructure_class,
@@ -62,8 +64,8 @@ class MultiprocessRunner(SystemRunner):
                     poll_interval=self.poll_interval,
                     pipeline_id=pipeline_id,
                     setup_tables=self.setup_tables,
-                    inbox=self.inboxes[(pipeline_id, process_name.lower())],
-                    outbox=self.outboxes[(pipeline_id, process_name.lower())],
+                    inbox=inbox,
+                    outbox=outbox,
                 )
                 os_process.daemon = True
                 os_process.start()
@@ -74,8 +76,9 @@ class MultiprocessRunner(SystemRunner):
 
     def broadcast_prompt(self, prompt):
         outbox_id = (prompt.pipeline_id, prompt.process_name)
-        assert outbox_id in self.outboxes, (outbox_id, self.outboxes.keys())
-        self.outboxes[outbox_id].put(prompt)
+        outbox = self.outboxes.get(outbox_id)
+        if outbox:
+            outbox.put(prompt)
 
     @staticmethod
     def is_prompt(event):
@@ -116,14 +119,15 @@ class OperatingSystemProcess(multiprocessing.Process):
         self.setup_tables = setup_tables
 
     def run(self):
-        # Construct process application object.
+        # Construct process application class.
         process_class = self.application_process_class
-        if self.infrastructure_class:
-            process_class = process_class.mixin(self.infrastructure_class)
+        if not isinstance(process_class, ApplicationWithConcreteInfrastructure):
+            if self.infrastructure_class:
+                process_class = process_class.mixin(self.infrastructure_class)
+            else:
+                raise ProgrammingError('infrastructure_class is not set')
 
-        if not issubclass(process_class, ApplicationWithConcreteInfrastructure):
-            raise Exception("Does not have application infrastructure: {}".format(process_class))
-
+        # Construct process application object.
         self.process = process_class(
             pipeline_id=self.pipeline_id,
             setup_table=self.setup_tables,
@@ -146,6 +150,7 @@ class OperatingSystemProcess(multiprocessing.Process):
                 # an API from which we can pull. It's not unreasonable to have a fixed
                 # number of application processes connecting to the same database.
                 record_manager = self.process.event_store.record_manager
+
                 notification_log = RecordManagerNotificationLog(
                     record_manager=record_manager.clone(
                         application_name=upstream_name,
@@ -212,7 +217,8 @@ class OperatingSystemProcess(multiprocessing.Process):
         self.process.run(prompt)
 
     def broadcast_prompt(self, prompt):
-        self.outbox.put(prompt)
+        if self.outbox is not None:
+            self.outbox.put(prompt)
 
     @staticmethod
     def is_prompt(event):