+++ eventsourcing/infrastructure/django/manager.py
@@ -1,22 +1,20 @@
 from django.db import IntegrityError, connection, transaction, ProgrammingError
 
 from eventsourcing.infrastructure.base import SQLRecordManager
-from eventsourcing.infrastructure.django.models import NotificationTrackingRecord
 
 
 class DjangoRecordManager(SQLRecordManager):
-    tracking_record_class = NotificationTrackingRecord
 
     _where_application_name_tmpl = (
         " WHERE application_name = %s AND pipeline_id = %s"
     )
 
-    def write_records(self, records, tracking_kwargs=None):
+    def write_records(self, records, tracking_kwargs=None, orm_objs=None):
         try:
             with transaction.atomic(self.record_class.objects.db):
                 with connection.cursor() as cursor:
+                    # Insert tracking record.
                     if tracking_kwargs:
-                        # Insert tracking record.
                         params = [tracking_kwargs[c] for c in self.tracking_record_field_names]
                         cursor.execute(self.insert_tracking_record, params)
 
@@ -61,6 +59,11 @@ class DjangoRecordManager(SQLRecordManager):
                         # Save record objects.
                         for record in records:
                             record.save()
+
+                    # Call 'save()' on each of the ORM objects.
+                    if orm_objs:
+                        for orm_obj in orm_objs:
+                            orm_obj.save()
 
         except IntegrityError as e:
             self.raise_record_integrity_error(e)